requirements.txt
absl-py==2.1.0
accelerate==0.33.0
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
airsim==1.8.1
albucore==0.0.20
albumentations==1.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
apptools==5.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
babel==2.16.0
beautifulsoup4==4.12.3
black==24.10.0
bleach==6.1.0
blinker==1.8.2
build==1.2.2.post1
catboost==1.2.7
category-encoders==2.6.4
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7
configobj==5.0.9
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.10
dash==2.18.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses-json==0.5.14
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
easydict==1.13
einops==0.8.0
envisage==7.0.3
eval_type_backport==0.2.0
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.115.2
fastjsonschema==2.20.0
ffmpy==0.4.0
filelock==3.15.4
Flask==3.0.3
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.3.1
fvcore==0.1.5.post20221221
gradio==5.0.2
gradio_client==1.4.0
graphviz==0.20.3
greenlet==3.0.3
grpcio==1.66.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.25.2
hydra-core==1.3.2
idna==3.7
imageio==2.36.0
importlib_metadata==8.0.0
intel-openmp==2021.4.0
iopath==0.1.10
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_core==5.7.2
jupyter_server_terminals==0.5.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
kornia==0.7.2
kornia_rs==0.1.5
langchain==0.2.11
langchain-community==0.2.10
langchain-core==0.2.23
langchain-openai==0.1.17
langchain-text-splitters==0.2.2
langsmith==0.1.93
lazy_loader==0.4
lightgbm==4.5.0
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib
matplotlib-inline==0.1.7
mayavi==4.8.2
mdurl==0.1.2
meson==1.6.1
mistune==3.0.2
mkl==2021.4.0
more-itertools==10.5.0
mpi==1.0.0
mpi4py==4.0.1
mpmath==1.3.0
msgpack==1.0.8
msgpack-rpc-python==0.4
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
notebook_shim==0.2.4
numba==0.60.0
numexpr==2.10.1
numpy==1.25.2
omegaconf==2.3.0
open3d==0.18.0
openai==1.61.0
openai-whisper==20240930
openapi-schema-pydantic==1.2.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
orjson==3.10.6
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathfinding==1.0.11
pathfinding3d==0.7.2
pathspec==0.12.1
patsy==1.0.1
Pillow==9.5.0
platformdirs==4.2.2
plotly==5.23.0
portalocker==2.10.1
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==5.28.2
psutil==6.0.0
pure_eval==0.2.3
pycocotools==2.0.8
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
pyface==8.0.0
Pygments==2.18.0
pyparsing==3.1.2
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-multipart==0.0.12
pytz==2024.2
PyWavelets==1.8.0
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.1.0
qtconsole==5.5.2
QtPy==2.4.1
qudida==0.0.4
referencing==0.35.1
regex==2024.5.15
replicate==1.0.4
requests==2.32.3
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.2
rpds-py==0.20.0
ruff==0.6.9
safetensors==0.4.3

scikit-image==0.21.0
scikit-learn==1.5.2
scipy==1.14.0
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
shellingham==1.5.4
simsimd==6.0.5
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
space-time-astar==0.8
SQLAlchemy==2.0.31
stack-data==0.6.3
starlette==0.39.2
statsmodels==0.14.4
stringzilla==3.10.10
supervision==0.21.0
sympy==1.13.0
tabulate==0.9.0
tbb==2021.13.1
tenacity==8.5.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tiktoken==0.7.0
timm==1.0.7
tinycss2==1.3.0
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.0
tornado
tqdm==4.66.4
traitlets==5.14.3
traits==6.4.3
traitsui==8.0.0
transformers==4.42.4
typer==0.12.5
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.2.2
uvicorn==0.31.1
vtk==9.3.1
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.4
widgetsnbextension==4.0.11
xlib==0.21
yacs==0.1.8
yapf==0.40.2
yarl==1.9.4
zipp==3.19.2
